debug cart, quantities now apply

+++ models/cart.go
@@ -13,14 +13,14 @@ type Cart struct {
 	CreatedAt     time.Time  `json:"created_at"`
 	LastUpdatedAt time.Time  `json:"last_updated_at"`
 	Items         []CartItem `json:"items"`
-	TotalValue    float64    `json:"total_value"`
+	TotalValue    float32    `json:"total_value"`
 }
 
 type CartItem struct {
 	ID         string              `json:"id"`
 	CartID     string              `json:"cart_id"`
 	Name       string              `json:"name"`
-	SalePrice  float64             `json:"sale_price"`
+	SalePrice  float32             `json:"sale_price"`
 	Components []CartItemComponent `json:"components"`
 	Qty        int                 `json:"qty"`
 	CreatedAt  time.Time           `json:"created_at"`
